Update scripts section to match latest experience
I am currently learning Next.JS and it looks like the default behavior is to create the structure and edit package.json for you when you run the command:

```bash
npm install next react react-dom
```

I have made a minor change to the wording to reflect this. Let me know if you would like to change it. Thanks!